backend/requirements.txt
fastapi
uvicorn
bcrypt
PyJWT
python-dotenv
